refactor(test): move test utils
